Weird stuff happens when stepSize is a very small value due to lack of precision that makes newRow = newHeight prematurely, so clamped stepSize to a minimum value of 0.05f.

+++ COMP371/Terrain.cpp
@@ -260,7 +260,7 @@ void Terrain::getCatMullXVertices(float stepSize)
 	setupMesh(false);
 }
 
-void Terrain::getCatMullZVertices(float stepSize)
+void Terrain::getCatMullZVertices(double stepSize)
 {
 
 	// store width and height of reduced vertices
